+++ NEWS.md
@@ -1,1 +1,11 @@
-### Version 1.0.0 - First commit (21/07/2016)
+### Version 1.1.0 (2016-08-15)
+
+Major update from initial version with the following changes:
+
+* The function for finding tickers in the ftp now looks for the closest date in the case that the actual date is missing from the ftp
+* The function for finding tickers now returns a dataframe with the tickers and number of trades
+* Added control for bad files
+* The output for raw and agg type of outuput were revised
+* The vignette is revised
+
+### Version 1.0.0 - First commit (2016-07-21)